feat(heater-shaker): print subdegree temps
We convert and hold all our temperatures in doubles, but weren't
printing them that way, because, well, look at this commit.

We now print them, which requires using snprintf, because nobody's
implemented std::to_chars for float or double.

+++ stm32-modules/heater-shaker/include/heater-shaker/host_comms_task.hpp
@@ -361,8 +361,7 @@ requires MessageQueue<QueueImpl<Message>, Message> class HostCommsTask {
                                           errors::ErrorCode::GCODE_CACHE_FULL));
         }
         auto message = messages::SetTemperatureMessage{
-            .id = id,
-            .target_temperature = static_cast<uint32_t>(gcode.temperature)};
+            .id = id, .target_temperature = gcode.temperature};
         if (!task_registry->heater->get_message_queue().try_send(
                 message, TICKS_TO_WAIT_ON_SEND)) {
             auto wrote_to = errors::write_into(
